FEAT: usuarios comuns podem inserir e deletar apenas suas proprias participações

+++ src/main/java/com/gs/ecoocean/service/ParticipacaoService.java
@@ -3,7 +3,9 @@ package com.gs.ecoocean.service;
 import com.gs.ecoocean.dto.participacao.ParticipacaoCreateDTO;
 import com.gs.ecoocean.model.Participacao;
 import com.gs.ecoocean.model.Partida;
+import com.gs.ecoocean.model.User;
 import com.gs.ecoocean.model.Voluntario;
+import com.gs.ecoocean.model.enuns.PerfilUsuario;
 import com.gs.ecoocean.model.enuns.StatusPartida;
 import com.gs.ecoocean.repository.ParticipacaoRepository;
 import jakarta.persistence.EntityNotFoundException;
@@ -34,11 +36,15 @@ public class ParticipacaoService {
     }
 
     public Participacao create(ParticipacaoCreateDTO participacaoCreateDTO){
+        User authenticatedUser = UserService.authenticated();
+        Voluntario voluntario = voluntarioService.get(participacaoCreateDTO.idVoluntario());
+        if (!authenticatedUser.hasRole(PerfilUsuario.ADMIN) && !authenticatedUser.getId().equals(voluntario.getAuth().getId()))
+            throw new IllegalArgumentException("Voluntarios não tem permissão para inserir outros voluntários em partidas");
+
         Partida partida = partidaService.get(participacaoCreateDTO.idPartida());
         if (!partida.getStatus().equals(StatusPartida.AGENDADA) && !partida.getStatus().equals(StatusPartida.ATIVA))
             throw new DataIntegrityViolationException("Não é possível inserir participações em partidas encerradas ou canceladas");
 
-        Voluntario voluntario = voluntarioService.get(participacaoCreateDTO.idVoluntario());
         if (existsByPartidaAndVoluntario(partida.getId(),voluntario.getId()))
             throw new DataIntegrityViolationException("Esse voluntário já está na partida");
 
@@ -48,6 +54,11 @@ public class ParticipacaoService {
     public void deleteById(Long id){
         Participacao participacao = repository.findById(id).orElseThrow(() -> new EntityNotFoundException(
                 "não foi possível deletar a participação, participação não encontrada para o id:"+id));
+
+        User authenticatedUser = UserService.authenticated();
+        if (!authenticatedUser.hasRole(PerfilUsuario.ADMIN) &&
+                !participacao.getVoluntario().getAuth().getId().equals(authenticatedUser.getId()))
+            throw new DataIntegrityViolationException("Você não tem permissão para remover esse voluntário da partida");
 
         repository.deleteById(id);
     }
